faceapi/requirements.txt
# Face API dependencies
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# InsightFace and dependencies
insightface==0.7.3
onnxruntime>=1.17.0  # Use onnxruntime-gpu for GPU support
numpy>=1.24.0,<2.0.0
opencv-python-headless==4.8.1.78
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
